refactor(compat): move the header to detail/
