+++ include/co_context/detail/task_info.hpp
@@ -1,6 +1,6 @@
 #pragma once
 #include "co_context/config.hpp"
-#include "co_context/compat.hpp"
+#include "co_context/detail/compat.hpp"
 #include "co_context/log/log.hpp"
 #include <coroutine>
 #include <memory>
