+++ lib/co_context/detail/worker_meta.cpp
@@ -1,7 +1,7 @@
 #include "co_context/config.hpp"
 #include "co_context/co/condition_variable.hpp"
 #include "co_context/co/semaphore.hpp"
-#include "co_context/compat.hpp"
+#include "co_context/detail/compat.hpp"
 #include "co_context/detail/task_info.hpp"
 #include "co_context/detail/thread_meta.hpp"
 #include "co_context/detail/user_data.hpp"
